+++ week5/Frontend/src/apis/auth/login.ts
@@ -1,8 +1,8 @@
 import type { ApiResponse } from "../../types/ApiResponse";
-import { handleApiError } from "../../shared/handleApiError";
-import { apiPublic } from "../../shared/axios";
+import { handleApiError } from "../../shared/apiConfig/handleApiError";
+import { apiPublic } from "../../shared/apiConfig/axios";
 import type { LoginPayload, LoginResponse } from "../../types/user";
-import { authStorage } from "../../shared/authStorage";
+import { authStorage } from "../../shared/apiConfig/authStorage";
 
 export const login = async (payload: LoginPayload): Promise<LoginResponse> => {
     try {
