+++ week5/Frontend/src/shared/apiConfig/tokenUtils.ts
@@ -1,8 +1,9 @@
 import type { Axios, AxiosRequestConfig } from "axios";
 import { authStorage } from "./authStorage";
 import { jwtDecode } from "jwt-decode";
-import { getRefreshTokens } from "../apis/auth/getRefreshTokens";
+import { getRefreshTokens } from "../../apis/auth/getRefreshTokens";
 
+// check if the access token is expired
 export function isTokenExpired(bufferSeconds:number = 5): boolean {
     const accessToken = authStorage.getToken('access');
     if (!accessToken) return true;
@@ -17,16 +18,19 @@ export function isTokenExpired(bufferSeconds:number = 5): boolean {
     return exp <= currentTime + bufferSeconds;
 }
 
+// refresh the access token using the refresh token
 let refreshPromise: Promise<void> | null = null;
 
 export async function refreshToken(): Promise<void> {
     const refreshTokenValue = authStorage.getToken('refresh');
 
+    // no refresh token, logout
     if (!refreshTokenValue) {
         authStorage.logout();
         return Promise.reject(new Error("No refresh token"));
     }
 
+    // go when there's no ongoing refresh
     if (!refreshPromise) {
         refreshPromise = (async () => {
             try {
